make some minor performance improvements
change some types of variables
remove unnecessary check_config method from request of network library

+++ include/network.h
@@ -93,7 +93,7 @@
 	unsigned long combine_bytes(unsigned char *bytes, unsigned long byte_count);
 	struct Header get_header(struct Header *headers, const unsigned long header_size, const char *name);
 	struct hostent *resolve_hostname(char *hostname);
-	void add_field_to_formdata(struct FormData *formdata, const char *name, const char *data, const unsigned long data_size, const char *filename);
+	void add_field_to_formdata(struct FormData *formdata, const char *name, const char *data, const unsigned long long data_size, const char *filename);
 	void add_header_to_formdata_field(struct FormData *formdata, const char *field_name, const char *header_name, const char *header_value);
 	void free_formdata(struct FormData formdata);
